+++ src/Lib.php
@@ -5,6 +5,8 @@
 namespace Ktomk\Pipelines;
 
 
+use Ktomk\Pipelines\File\ImageName;
+
 class Lib
 {
     /**
@@ -181,28 +183,5 @@ class Lib
             }
         }
         return 0;
-    }
-
-    /**
-     * Is a Docker image name (optionally with a tag) syntactically
-     * valid?
-     *
-     * @see doc/DOCKER-NAME-TAG.md
-     *
-     * @param string $name of docker image
-     * @return bool
-     */
-    public static function validDockerImage($name)
-    {
-        $pattern =
-            '{^' .
-                '([a-zA-Z0-9.-]+(:[0-9]+)?/)?' . # <prefix>
-                '([a-z0-9]+(?:(?:\.|__?|-+)[a-z0-9]+)*)(/[a-z0-9]+(?:(?:\.|__?|-+)[a-z0-9]+)*)*' . # <name-components>
-                '(:[a-zA-Z0-9_][a-zA-Z0-9_.-]{0,127})?' . # <tag-name>
-            '$}';
-
-        $result = preg_match($pattern, $name);
-
-        return 1 === $result;
     }
 }