+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+### Fixed
+- Image name validation
+- Image as a section
+
 ## [0.0.4] - 2018-01-16
 ### Added
 - Release phar files on Github